Changed check for target file to avoid tripping up test (which moves files to same location)

+++ src/1. Layers/1.5 ACL/PH.Well.Task.GlobalUplifts/Csv/CsvUpliftDataProvider.cs
@@ -169,7 +169,7 @@ namespace PH.Well.Task.GlobalUplifts.Csv
             if (_archiveFile)
             {
                 var target = Path.Combine(_archivePath, _id);
-                if (File.Exists(target))
+                if (File.Exists(target) && target != _filePath)
                 {
                     File.Delete(target);
                 }
